fix: use long for bot/user IDs, broaden delete rollback, fix delete order
Agent-Logs-Url: https://github.com/iAlexX11/Project_S2_OOPD/sessions/4b732341-94df-447b-8c40-bac5dfd82a8c

+++ src/main/java/org/cryptoBros/persistence/AtomicPersistence.java
@@ -8,7 +8,7 @@ import org.cryptoBros.persistence.Exceptions.DbConnectionException;
 
 public interface AtomicPersistence {
 
-    int createCryptoWithBot(Crypto crypto)
+    long createCryptoWithBot(Crypto crypto)
             throws DbConnectionException, BotGenerationException, CryptoNotAddedException;
 
     void deleteCryptoWithBot(String symbol)
